Added README.
Bugfix in the makefiles/scripts. Moved the file copy between the yahtzee make and the unit tests to its own script, because makefiles evaluate expressions at the beginning of execution instead of in order, so newly copied files weren't picked up.

Updated unit test includes to reflect that files are now copied to the local directory instead of trying to link with yahtzee libraries.

Updated the top level yahtzee function to give better and cleaner print output.